requirements.txt
Flask==2.0.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
google-cloud-storage==1.42.0
requests==2.26.0
python-dotenv==0.19.0
apscheduler==3.8.1
Werkzeug==2.0.1
psycopg2-binary==2.9.3